companyquery methods first round of testing

+++ db/query.js
@@ -1,6 +1,7 @@
 ///create a class (Query) that will contain the methods that view/create/update/delete data
 
 const { Pool } = require("pg");
+// const prompts = require("../directions/prompts.js");
 require("dotenv").config();
 
 const pool = new Pool({
@@ -15,21 +16,33 @@ pool.connect();
 class CompanyQuery {
   //View all Departments (names, deparment_ids): viewDepartments()
   viewDepartments() {
-    pool.query("SELECT * FROM departments", function (err, { rows }) {
+    pool.query("SELECT * FROM departments", (err, { rows }) => {
+      if (err) {
+        console.log(err);
+        return;
+      }
       console.table(rows);
     });
   }
 
   //View all Roles (Job Title, role id, Department(name), salary): viewRoles()
   viewRoles() {
-    pool.query("SELECT * FROM roles", function (err, { rows }) {
+    pool.query("SELECT * FROM roles", (err, { rows }) => {
+      if (err) {
+        console.log(err);
+        return;
+      }
       console.table(rows);
     });
   }
 
   //View all Employees (employee id, fist name, last name, job title, department, salary, manager name): viewEmployees()
   viewEmployees() {
-    pool.query("SELECT * FROM employees", function (err, { rows }) {
+    pool.query("SELECT * FROM employees", (err, { rows }) => {
+      if (err) {
+        console.log(err);
+        return;
+      }
       console.table(rows);
     });
   }
@@ -37,11 +50,14 @@ class CompanyQuery {
   //Add Department (provie name): addDepartment(name): addDepartment
   addDepartment(departmentName) {
     pool.query(
-      `INSERT INTO departments (name)
-    VALUE ($1)`,
+      "INSERT INTO departments (name)  VALUES ($1)",
       [departmentName],
-      function (err, { rows }) {
-        console.table(rows);
+      (err, result) => {
+        if (err) {
+          console.log(err);
+          return;
+        }
+        console.log("Success!");
       }
     );
   }
@@ -49,11 +65,14 @@ class CompanyQuery {
   //Add Role (provide name, salary and department): addRole()
   addRole(roleTitle, roleSalary, roleDepartmentId) {
     pool.query(
-      `INSERT INTO roles (title, salary, department_id)
-    VALUE ($1, $2, $3)`,
+      "INSERT INTO roles (title, salary, department_id)  VALUES ($1, $2, $3)",
       [roleTitle, roleSalary, roleDepartmentId],
-      function (err, { rows }) {
-        console.table(rows);
+      (err, result) => {
+        if (err) {
+          console.log(err);
+          return;
+        }
+        console.log("Success!");
       }
     );
   }
@@ -66,78 +85,90 @@ class CompanyQuery {
     employeeManagerId
   ) {
     pool.query(
-      `INSERT INTO roles (first_name, last_name, role_id, manager_id)
-    VALUE ($1, $2, $3, $4)`,
+      "INSERT INTO employees (first_name, last_name, role_id, manager_id)  VALUES ($1, $2, $3, $4)",
       [employeeFirstName, employeeLastName, employeeRoleId, employeeManagerId],
-      function (err, { rows }) {
-        console.table(rows);
+      (err, result) => {
+        if (err) {
+          console.log(err);
+          return;
+        }
+        console.log("Success!");
       }
     );
   }
   //Update Employee (select employee provide field to update and updatevalue): updateEmployee()
   updateEmployee(updateField, updateValue, employeeId) {
-    pool.query(
-      `UPDATE employees
-    SET $1 = $2
-    WHERE id = $3`,
-      [updateField, updateValue, employeeId],
-      function (err, { rows }) {
-        console.table(rows);
+    const sql = `UPDATE employees  SET ${updateField} = $1  WHERE id = $2`;
+
+    pool.query(sql, [updateValue, employeeId], (err, result) => {
+      if (err) {
+        console.log(err);
+        return;
       }
-    );
+      console.log("Success!");
+    });
   }
   //Update Role: updateRole()
   updateRole(updateField, updateValue, roleId) {
-    pool.query(
-      `UPDATE roles
-    SET $1 = $2
-    WHERE id = $3`,
-      [updateField, updateValue, roleId],
-      function (err, { rows }) {
-        console.table(rows);
+    const sql = `UPDATE roles  SET ${updateField} = $1  WHERE id = $2`;
+
+    pool.query(sql, [updateValue, roleId], (err, result) => {
+      if (err) {
+        console.log(err);
+        return;
       }
-    );
+      console.log("Success!");
+    });
   }
   //Update Manager: updateManager()
   updateManager(managerId, employeeId) {
-    pool.query(
-      `UPDATE employees
-    SET manager_id = $1
-    WHERE id = $2`,
-      [managerId, employeeId],
-      function (err, { rows }) {
-        console.table(rows);
+    const sql = `UPDATE employees  SET manager_id = $1  WHERE id = $2`;
+
+    pool.query(sql, [managerId, employeeId], (err, result) => {
+      if (err) {
+        console.log(err);
+        return;
       }
-    );
+      console.log("Success!");
+    });
   }
   //Update Department: updateDepartment()
   updateDepartment(updateValue, departmentId) {
-    pool.query(
-      `UPDATE departments
-    SET name = $1
-    WHERE id = $2`,
-      [updateValue, departmentId],
-      function (err, { rows }) {
-        console.table(rows);
+    const sql = `UPDATE departments  SET name = $1  WHERE id = $2`;
+
+    pool.query(sql, [updateValue, departmentId], (err, result) => {
+      if (err) {
+        console.log(err);
+        return;
       }
-    );
+      console.log("Success!");
+    });
   }
   //Manager's Employees: managerEmployees()
   managerEmployees(managerId) {
     pool.query(
-      `SELECT * FROM employees WHERE manager_id = $1`,
+      "SELECT * FROM employees  WHERE manager_id = $1",
       [managerId],
-      function (err, { rows }) {
+      (err, { rows }) => {
+        if (err) {
+          console.log(err);
+          return;
+        }
         console.table(rows);
       }
     );
   }
   //Department Employees: departmentEmployees()
   departmentEmployees(employeeDepartmentId) {
+    //departments_id comes from the role of the employee
     pool.query(
-      `SELECT * FROM employees WHERE department_id = $1`,
+      "SELECT roles.department_id AS Department_Number, employees.first_name AS First_Name, employees.last_name AS Last_Name, employees.id AS Employee_ID, employees.manager_id AS Manager_Employee_Id FROM roles JOIN employees ON roles.id = employees.role_id  WHERE roles.department_id = $1",
       [employeeDepartmentId],
-      function (err, { rows }) {
+      (err, { rows }) => {
+        if (err) {
+          console.log(err);
+          return;
+        }
         console.table(rows);
       }
     );
@@ -147,36 +178,48 @@ class CompanyQuery {
     pool.query(
       `DELETE FROM departments WHERE id = $1`,
       [departmentId],
-      function (err, { rows }) {
-        console.table(rows);
+      (err, results) => {
+        if (err) {
+          console.log(err);
+          return;
+        }
+        console.log("DELETED");
       }
     );
   }
   //Delete Roles: deleteRole()
   deleteRole(roleId) {
-    pool.query(
-      `DELETE FROM roles WHERE id = $1`,
-      [roleId],
-      function (err, { rows }) {
-        console.table(rows);
+    pool.query(`DELETE FROM roles WHERE id = $1`, [roleId], (err, results) => {
+      if (err) {
+        console.log(err);
+        return;
       }
-    );
+      console.log("DELETED");
+    });
   }
   //Delete Employees: deleteEmployee()
   deleteEmployee(employeeId) {
     pool.query(
       `DELETE FROM employees WHERE id = $1`,
       [employeeId],
-      function (err, { rows }) {
-        console.table(rows);
+      (err, results) => {
+        if (err) {
+          console.log(err);
+          return;
+        }
+        console.log("DELETED");
       }
     );
   }
   //DepartmentUtilizedBudget (sum of employee salaries): utilizedBudget()
   utilizedBudget() {
     pool.query(
-      `SELECT departments.name AS department, roles.department_id AS department_number, SUM(roles.salary) AS utilized_budget FROM roles GROUP BY roles.department_id JOIN departments ON roles.department_id = departments.id`,
-      function (err, { rows }) {
+      `SELECT SUM(roles.salary) AS utilized_budget, roles.department_id AS department_number FROM roles GROUP BY roles.department_id `,
+      (err, { rows }) => {
+        if (err) {
+          console.log(err);
+          return;
+        }
         console.table(rows);
       }
     );
